+++ importers/src/bible/sword/src/VerseMetadata.ts
@@ -1,5 +1,5 @@
 import VerseScheme from './VerseScheme';
-const bcvParser = require('bible-passage-reference-parser/js/en_bcv_parser').bcv_parser;
+const bcvParser = require('bible-passage-reference-parser/cjs/en_bcv_parser').bcv_parser;
 
 const bcv = new bcvParser();